requirements.txt
openai
tiktoken
pypdf
python-docx
gradio==4.21.0
python-dotenv
pydantic==2.8.2
fastapi==0.112.4
